Refactor Node and Value Implementations; Enhance Plugin API
- Updated the `Node` struct to remove unnecessary lifetime parameters and associated implementations.
- Refactored the `Value` and `ValueNotArray` traits to simplify their implementations for `Node`.
- Introduced new methods in the `Node` struct for frame caching and retrieval.
- Enhanced the `Plugin` API with new methods for configuration and function registration.
- Added tests for the `ScriptAPI` to verify API version compatibility.
- Updated the `Environment` methods to use `i32` for indices instead of `usize`.
- Improved error handling and string conversions in plugin methods.

+++ rustsynth-sys/src/lib.rs
@@ -12,9 +12,9 @@ include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
 
 macro_rules! version {
     ($major:expr, $minor:expr) => {
-        (($major) << 16) | ($minor)
+        (($major) << 16) | ($minor) 
     };
 }
 
-pub const VAPOURSYNTH_API_VERSION: u32 = version!(VAPOURSYNTH_API_MAJOR, VAPOURSYNTH_API_MINOR);
-pub const VSSCRIPT_API_VERSION: u32 = version!(VSSCRIPT_API_MAJOR, VSSCRIPT_API_MINOR);
+pub const VAPOURSYNTH_API_VERSION: i32 = version!(VAPOURSYNTH_API_MAJOR as i32, VAPOURSYNTH_API_MINOR as i32);
+pub const VSSCRIPT_API_VERSION: i32 = version!(VSSCRIPT_API_MAJOR as i32, VSSCRIPT_API_MINOR as i32);
